GossipGraph

Documentation

data Kind

Edge label indicating whether an agent knows the number or secret of another agent

Constructors

Number

Secret

type Relation = LEdge Kind

type GossipGraph = Gr AgentName Kind

printGraph :: GossipGraph -> IO ()

testGraph :: GossipGraph

Simple graph to be used for testing

testGraph2 :: GossipGraph

biggerGraph :: GossipGraph

defaultGraph :: Int -> GossipGraph

initialGraph :: Int -> [(Char, [Char])] -> GossipGraph

Generate an initial gossip graph (with no initial shared secrets), based on a list of agents and their known phone numbers.

idToLab :: Int -> Char

Convert an agent ID to an agent label

labToId :: Char -> Int

Convert an agent label to an agnet ID

hasRelationWith :: GossipGraph -> Agent -> Kind -> Agent -> Bool

numbersKnownBy :: GossipGraph -> Agent -> [Agent]

secretsKnownBy :: GossipGraph -> Agent -> [Agent]

noAgents :: GossipGraph -> Int

agent :: Int -> Char -> Agent

Warning, ignores the Char argument! Remains for legacy purposes.

agentFromId :: Int -> Agent

agentFromLab :: Char -> Agent

relation :: Agent -> Agent -> Kind -> Relation

findAgentByName :: [Agent] -> AgentName -> Maybe Agent

Given a list of agents and a name, try to find a matching agent

graphIsComplete :: GossipGraph -> Bool

Check whether each agent is an expert; i.e. knows the secret of everyone.

data LexToken

Possible tokens in the input string

Constructors

Token Kind AgentName AgentId

Separator

lexer :: String -> Maybe [LexToken]

Lexing for gossip graph input

parseAgents :: [LexToken] -> Maybe [Agent]

Extract the agents from the lexed input string

parseRelations :: [Agent] -> [LexToken] -> Maybe [Relation]

Extract the relations from the parsed agents and the lexed input string

fromString :: String -> Maybe (Gr AgentName Kind)

Try to parse a string representation of a gossip graph into a gossip graph

fromAgentsAndRelations :: [Agent] -> [Relation] -> Gr AgentName Kind

Given a set of agents and corresponding relations, construct a gossip graph

toStringIfSuccessful :: Maybe (Gr Agent Relation) -> String

Debugging function to check if parsing was successful