GossipState

State type

data State

The gossip state.

Constructors

State

Fields

stateGraph :: GossipGraph

The gossip graph of the current state.

stateKnowledgeStruct :: GossipKnowledgeStructure

The knowledge structure of the current state.

stateCallSeq :: [Call]

The call sequence of the current state.

printState :: State -> Bool -> IO ()

Prints the current state.

Call validation and execution

validCalls :: GossipGraph -> ([Call], [GroupCall])

Determines based on the current GossipGraph state which calls are actually allowed to be made. Note that no protocols are used for this, a call between a and b is valid iff a knows the number of b.

makeCall :: Call -> State -> State

Makes a call and update the state.

executeCalls :: [Call] -> State -> IO State

Makes multiple calls by repeatedly invoking makeCall.

Epistemic formula evaluation

evaluateGossipAtom :: State -> GossipAtom -> Bool

Evaluates a gossip atom (N(x,y), S(x,y) or C(x,y)), given the current state

evaluateBddVar :: State -> Int -> Bool

Evaluates a Bdd variable as Int, given the current state

evaluate :: State -> Form -> Bool

Evaluates an epistemic formula, given the current state

(|=) :: State -> Form -> Bool infix 9

An infix operator of the evaluate function.

Legacy

evaluate' :: State -> GossipForm -> Bool

Legacy method, evalute a GossipFormula (note, without knowledge) by naive recursion