GossipTypes

Documentation

type Agent = LNode Char

type AgentId = Int

type AgentName = Char

type Call = (Agent, Agent)

(from, to). Call is considered to be bidirectional, Optionally we could add a direction component to allow singe directional broadcasting calls.

showAgent :: Agent -> String

printCall :: Call -> IO ()

printMakeCall :: Call -> IO ()

printCalls :: [Call] -> IO ()