Util

Description

Coloured printing

putStrFgc :: Color -> String -> IO ()

Prints the given message in a given colour.

putStrLnFgc :: Color -> String -> IO ()

Prints the given mesage with a new line in a given colour.

Messaging

printInvalidAction :: Char -> IO ()

Prints a message that an action is invalid for the current context.

printGraphComplete :: IO ()

Prints a message that the current graph is complete, i.e. every agent is an expert.

Pretty BDD to graph saving

writeToSvgGraph :: String -> Bdd -> IO ()

Save a Bdd in its graph structure as an SVG file. Uses the dot language, for which Graphviz needs to be installed.