GossipGraph

Description

Gossip Graph types

type GossipGraph = Gr AgentName Kind

The gossip graph. This is defined in terms of the Graph module.

data Kind

An agent relation label, indicating whether x knows the number of y, or x knows the secret of y.

Constructors

Number

Secret

type Relation = LEdge Kind

A relation between to agents, either knowledge of number or knowledge of secret.

Pre-made graphs

testGraph :: GossipGraph

Simple graph to be used for testing

testGraph2 :: GossipGraph

Another simple graph with slightly different number relations.

biggerGraph :: GossipGraph

A slightly bigger graph, with five instead of three agents.

defaultGraph :: Int -> GossipGraph

A default graph with generic size. In this graph, every agent only knows their own number.

>>> defaultGraph 3
mkGraph [(0,'a'),(1,'b'),(2,'c')] [(0,0,Number),(0,0,Secret),(1,1,Number),(1,1,Secret),(2,2,Number),(2,2,Secret)]

Graph construction

initialGraph :: Int -> [(Char, [Char])] -> GossipGraph

Generates an initial gossip graph (with no initial shared secrets), based on a list of agents and their known phone numbers. In this initial graph, everyone will only know their own secret.

>>> initialGraph 2 [('a', "ab"), ('b', "b")]
mkGraph [(0,'a'),(1,'b')] [(0,0,Number),(0,0,Secret),(0,1,Number),(1,1,Number),(1,1,Secret)]

Graph inspection

numbersKnownBy :: GossipGraph -> Agent -> [Agent]

Returns the list of agents of which an agent knows the number, given a gossip graph.

secretsKnownBy :: GossipGraph -> Agent -> [Agent]

Returns the list of agents of which an agent knows the secret, given a gossip graph.

hasRelationWith :: GossipGraph -> Agent -> Kind -> Agent -> Bool

Checks whether two agents have some relation, either secret or number.

isGraphComplete :: GossipGraph -> Bool

Check whether each agent is an expert; i.e. knows the secret of everyone.

noAgents :: GossipGraph -> Int

Returns the amount of agents that are present in a gossip graph.

>>> noAgents testGraph
3

Agent-specific functions

idToLab :: Int -> Char

Converts an agent ID to an agent label.

>>> idToLab 0
'a'
>>> idToLab 12
'm'

labToId :: Char -> Int

Converts an agent label to an agent ID.

>>> labToId 'a'
0
>>> labToId 'w'
22

Agent construction

agentFromId :: Int -> Agent

Generates an agent, based on its ID.

>>> agentFromId 0
(0, 'a')
>>> agentFromId 2
(2, 'c')

agentFromLab :: Char -> Agent

Generates an agent, based on its label character.

>>> agentFromLab 'a'
(0, 'a')
>>> agentFromLab 'd'
(3, 'd')

Relation construction

relation :: Agent -> Agent -> Kind -> Relation

Generates a relation between to agents, given a relation kind.

Graph printing

printGraph :: GossipGraph -> IO ()

Prints the graph in a readable manner.