GossipKnowledge

Documentation

data Rel

The relation type of a GossipAtom.

Constructors

N
x knows the number of y

S
x knows the secret of y

C
x has called y

data GossipAtom

An atomic proposition in a gossip state: Rel(Agent, Agent)

Constructors

GAt Rel Agent Agent

texify :: GossipAtom -> String

bddToGAt :: Int -> Bdd -> [GossipAtom]

Converts Bdd formula to a list of GossipAtoms for all variables in the formula

varToGAt :: Int -> Int -> GossipAtom

Converts a Bbd variable index to a GossipAtom

showBddVar :: Int -> Int -> String

Convert a bdd variable index to a GossipAtom string

texifyBddVar :: Int -> Int -> String

strLabel :: Int -> VarLabeller

texLabel :: Int -> VarLabeller

gvar :: Int -> Int -> Bdd

gAtToBdd :: Int -> GossipAtom -> Bdd

Convert a GossipAtom to a Bdd variable

gAtToInt :: Int -> GossipAtom -> Int

Convert a GossipAtom to a unique integer

data GossipForm

A formula build o' of GossipAtoms, using the language of (van Ditmarsch et al., 2017). We're not certain we need this, we might fully stick with Bdd formulae instead.

Constructors

Top
Always true

Atom GossipAtom
Atom

Neg GossipForm
Negation

Conj [GossipForm]
Conjunction

Disj [GossipForm]
Disjunction

Impl GossipForm GossipForm
Implication

data Form

An epistemic formula, defined in terms of Bdd's

Constructors

Fact Bdd

K Agent Form

data GossipKnowledgeStructure

A knowledge structure (Gattinger, 2018) which represents the knowledge of a Gossip State

Constructors

GKS

Fields

vocabulary :: Set Int

The set of atoms available in the model

stateLaw :: Bdd

A boolean formula representing the law that every state needs to adhere to

observables :: Map Agent (Set Int)

The set of atoms seen by some agent

nag :: GossipKnowledgeStructure -> Int

validKS :: GossipKnowledgeStructure -> Bool

gforall :: GossipKnowledgeStructure -> Int -> Bdd -> Bdd

gexists :: GossipKnowledgeStructure -> Int -> Bdd -> Bdd

gforallSet :: GossipKnowledgeStructure -> [Int] -> Bdd -> Bdd

gexistsSet :: GossipKnowledgeStructure -> [Int] -> Bdd -> Bdd

fromGossipGraph :: GossipGraph -> GossipKnowledgeStructure

Converts a gossip graph to its corresponding knowledge structure. Note that this doesn't convert a gossip state, i.e. no call sequence is encoded. This is meant for the conversion of the _initial_ gossip state, (G, ε) (van Ditmarsch et al., 2017)

formToBdd :: GossipKnowledgeStructure -> Form -> Bdd

Convert an epistemic formula to a boolean formula, given a knowledge structure in Gattinger (2018), this is denoted by ||ϕ||_F

(<|>) :: GossipKnowledgeStructure -> Form -> Bdd infix 9

data KnowledgeTransformer

Constructors

KT

Fields

addVocab :: Set Int

Additional vocabulary

eventLaw :: Bdd

Event law

addObs :: Map Agent (Set Int)

Additional observables

baseTransformer :: KnowledgeTransformer

validKT :: GossipKnowledgeStructure -> KnowledgeTransformer -> Bool

update :: GossipKnowledgeStructure -> KnowledgeTransformer -> GossipKnowledgeStructure

(|+|) :: GossipKnowledgeStructure -> KnowledgeTransformer -> GossipKnowledgeStructure infixl 9

synchronousUpdate :: GossipKnowledgeStructure -> Int -> Call -> GossipKnowledgeStructure