GossipTypes

Documentation

type Agent = LNode Char

type AgentId = Int

type AgentName = Char

type Call = (Agent, Agent)

(from, to). Call is considered to be bidirectional, Optionally we could add a direction component to allow singe directional broadcasting calls.

type GroupCall = (Agent, [Agent])

(☎) :: Agent -> Agent -> Call infix 0

toCalls :: GroupCall -> [Call]

showAgent :: Agent -> String

printNoCalls :: IO ()

printCall :: Call -> IO ()

printMakeCall :: Call -> IO ()

printCalls :: [Call] -> IO ()

printGroupCall :: GroupCall -> IO ()

printMakeGroupCall :: GroupCall -> IO ()

printGroupCalls :: [GroupCall] -> IO ()

printAllCalls :: ([Call], [GroupCall]) -> IO ()